Harmonize wind direction data model + use carinal wind direction in all occasions

+++ OpenWeatherMap/Models/CurrentWeatherForecast.cs
@@ -64,7 +64,7 @@ namespace OpenWeatherMap.Models
         ///  Wind direction, degrees (meteorological).
         /// </summary>
         [JsonProperty("wind_deg")]
-        public int WindDirectionDegrees { get; set; }
+        public double WindDirectionDegrees { get; set; }
 
         /// <summary>
         ///  Cardinal wind direction.
